Fix channel type of NotificationFailed events
This commit changes the channel property that is set on the dispatched
NotificationFailed events from a class instance to a string representing
the class name. This is done because the property is documented as a
string property on Illuminate's side, and if an application assumes that
this package adheres to that documentation, runtime errors may occur.